tests/wpt/web-platform-tests/offscreen-canvas/line-styles/2d.line.miter.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.line.miter.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.line.miter.valid</h1>
<p class="desc">Setting miterLimit to valid values works</p>


<script>
var t = async_test("Setting miterLimit to valid values works");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.miterLimit = 1.5;
_assertSame(ctx.miterLimit, 1.5, "ctx.miterLimit", "1.5");
ctx.miterLimit = "1e1";
_assertSame(ctx.miterLimit, 10, "ctx.miterLimit", "10");
ctx.miterLimit = 1/1024;
_assertSame(ctx.miterLimit, 1/1024, "ctx.miterLimit", "1/1024");
ctx.miterLimit = 1000;
_assertSame(ctx.miterLimit, 1000, "ctx.miterLimit", "1000");

t.done();

});
</script>
